use unsigned short for number of tracks

+++ StEventPlane/StEventPlane.cxx
@@ -223,7 +223,7 @@ int StEventPlane::calculateEventPlane()
 
    // pre-loop to count tracks for event plane, prepare for shuffle
    int nTracksForEventPlane = 0;
-   for (int iTrack = 0; iTrack < mPicoDst->numberOfTracks(); ++iTrack)
+   for (unsigned short iTrack = 0; iTrack < mPicoDst->numberOfTracks(); ++iTrack)
    {
       StPicoTrack* picoTrack = (StPicoTrack*) mPicoDst->track(iTrack);
       if (!picoTrack)
@@ -262,7 +262,7 @@ int StEventPlane::calculateEventPlane()
    memset(QxEta, 0, 20*sizeof(float));
    memset(QyEta, 0, 20*sizeof(float));
    float vertexZ = mVertexPos.z();
-   for (int iTrack = 0; iTrack < mPicoDst->numberOfTracks(); ++iTrack)
+   for (unsigned short iTrack = 0; iTrack < mPicoDst->numberOfTracks(); ++iTrack)
    {
       StPicoTrack* picoTrack = (StPicoTrack*) mPicoDst->track(iTrack);
       if (!picoTrack)
@@ -427,7 +427,7 @@ TVector2 StEventPlane::QEtaGap(int iEta, int nEtaGaps) const
 
 void StEventPlane::calculateHadronV2() const
 {
-   for (int iTrack = 0; iTrack < mPicoDst->numberOfTracks(); iTrack++)
+   for (unsigned short iTrack = 0; iTrack < mPicoDst->numberOfTracks(); iTrack++)
    {
       StPicoTrack* picoTrack = (StPicoTrack*) mPicoDst->track(iTrack);
       if (!picoTrack)
